+++ resources/js/utils/products.js
@@ -1,5 +1,4 @@
 import { API_URL } from "../data";
-import { useUserStore } from "../stores/user";
 
 export const getProducts = async ({
     offset = 0,
@@ -20,18 +19,5 @@ export const getCategories = async () => {
 
 export const getSuppliers = async () => {
     const response = await fetch(`${API_URL}/api/suppliers`);
-    return response.json();
-};
-
-export const saveOrder = async (data) => {
-    const userStore = useUserStore();
-    const response = await fetch(`${API_URL}/api/orders`, {
-        method: "POST",
-        headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${userStore.token}`,
-        },
-        body: JSON.stringify(data),
-    });
     return response.json();
 };
